Change UE ID from uint8 to int64 trying to fix crash with Open5GS

+++ internal/control_test_engine/ue/nas/message/nas_control/mm_5gs/ul-nas-transport.go
@@ -41,7 +41,7 @@ func getUlNasTransport_PduSessionEstablishmentRequest(pduSessionId int64, reques
 	ulNasTransport.ExtendedProtocolDiscriminator.SetExtendedProtocolDiscriminator(nasMessage.Epd5GSMobilityManagementMessage)
 	ulNasTransport.PduSessionID2Value = new(nasType.PduSessionID2Value)
 	ulNasTransport.PduSessionID2Value.SetIei(nasMessage.ULNASTransportPduSessionID2ValueType)
-	ulNasTransport.PduSessionID2Value.SetPduSessionID2Value(int64(pduSessionId))
+	ulNasTransport.PduSessionID2Value.SetPduSessionID2Value(uint8(pduSessionId))
 	ulNasTransport.RequestType = new(nasType.RequestType)
 	ulNasTransport.RequestType.SetIei(nasMessage.ULNASTransportRequestTypeType)
 	ulNasTransport.RequestType.SetRequestTypeValue(requestType)
